refactor: move metrics to crate wide struct
