+++ src/influxdb_ioxd/http.rs
@@ -20,8 +20,8 @@ use data_types::{
 };
 use influxdb_iox_client::format::QueryOutputFormat;
 use influxdb_line_protocol::parse_lines;
+use metrics::IOXD_METRICS;
 use object_store::ObjectStoreApi;
-use once_cell::sync::Lazy;
 use query::{frontend::sql::SQLQueryPlanner, Database, DatabaseStore};
 use server::{ConnectionManager, Server as AppServer};
 
@@ -31,7 +31,7 @@ use futures::{self, StreamExt};
 use http::header::{CONTENT_ENCODING, CONTENT_TYPE};
 use hyper::{Body, Method, Request, Response, StatusCode};
 use observability_deps::{
-    opentelemetry::{metrics::Counter, KeyValue},
+    opentelemetry::KeyValue,
     tracing::{self, debug, error, info},
 };
 use routerify::{prelude::*, Middleware, RequestInfo, Router, RouterError, RouterService};
@@ -302,39 +302,6 @@ impl ApplicationError {
     }
 }
 
-/// Contains module-wide metrics
-struct HttpMetrics {
-    /// How many line protocol lines were parsed but rejected
-    lp_lines_errors: Counter<u64>,
-    /// How many line protocol lines were parsed and successfully loaded
-    lp_lines_success: Counter<u64>,
-    /// How many bytes of protocol line were parsed and successfully loaded
-    lp_bytes_success: Counter<u64>,
-}
-
-static METRICS: Lazy<HttpMetrics> = Lazy::new(HttpMetrics::new);
-
-impl HttpMetrics {
-    fn new() -> Self {
-        Self {
-            lp_lines_errors: metrics::meter()
-                .u64_counter("ingest.lp.lines.errors")
-                .with_description("line protocol formatted lines which were rejected")
-                .init(),
-
-            lp_lines_success: metrics::meter()
-                .u64_counter("ingest.lp.lines.success")
-                .with_description("line protocol formatted lines which were successfully loaded")
-                .init(),
-
-            lp_bytes_success: metrics::meter()
-                .u64_counter("ingest.lp.bytes.success")
-                .with_description("line protocol formatted bytes which were successfully loaded")
-                .init(),
-        }
-    }
-}
-
 const MAX_SIZE: usize = 10_485_760; // max write request size of 10MB
 
 fn router<M>(server: Arc<AppServer<M>>) -> Router<Body, ApplicationError>
@@ -483,7 +450,9 @@ where
     ];
 
     server.write_lines(&db_name, &lines).await.map_err(|e| {
-        METRICS.lp_lines_errors.add(lines.len() as u64, &metric_kv);
+        IOXD_METRICS
+            .lp_lines_errors
+            .add(lines.len() as u64, &metric_kv);
 
         match e {
             server::Error::DatabaseNotFound { .. } => ApplicationError::DatabaseNotFound {
@@ -497,8 +466,12 @@ where
         }
     })?;
 
-    METRICS.lp_lines_success.add(lines.len() as u64, &metric_kv);
-    METRICS.lp_bytes_success.add(body.len() as u64, &metric_kv);
+    IOXD_METRICS
+        .lp_lines_success
+        .add(lines.len() as u64, &metric_kv);
+    IOXD_METRICS
+        .lp_bytes_success
+        .add(body.len() as u64, &metric_kv);
 
     Ok(Response::builder()
         .status(StatusCode::NO_CONTENT)
